Move fake protocol to test utils
