+++ codegen/smithy-ruby-codegen-test-utils/src/main/java/software/amazon/smithy/ruby/codegen/integrations/WhiteLabelTestIntegration.java
@@ -15,16 +15,19 @@
 
 package software.amazon.smithy.ruby.codegen.integrations;
 
+import java.util.Arrays;
 import java.util.List;
 import software.amazon.smithy.model.Model;
 import software.amazon.smithy.model.shapes.ServiceShape;
 import software.amazon.smithy.ruby.codegen.GenerationContext;
+import software.amazon.smithy.ruby.codegen.ProtocolGenerator;
 import software.amazon.smithy.ruby.codegen.RubyIntegration;
 import software.amazon.smithy.ruby.codegen.RubyRuntimePlugin;
 import software.amazon.smithy.ruby.codegen.config.ClientConfig;
 import software.amazon.smithy.ruby.codegen.middleware.Middleware;
 import software.amazon.smithy.ruby.codegen.middleware.MiddlewareBuilder;
 import software.amazon.smithy.ruby.codegen.middleware.MiddlewareStackStep;
+import software.amazon.smithy.ruby.codegen.protocol.FakeProtocolGenerator;
 
 public class WhiteLabelTestIntegration implements RubyIntegration {
 
@@ -89,4 +92,8 @@ public class WhiteLabelTestIntegration implements RubyIntegration {
         middlewareBuilder.register(midMiddleware);
     }
 
+    @Override
+    public List<ProtocolGenerator> getProtocolGenerators() {
+        return Arrays.asList(new FakeProtocolGenerator());
+    }
 }
