+++ codegen/smithy-ruby-codegen-test-utils/src/main/resources/META-INF/services/software.amazon.smithy.ruby.codegen.RubyIntegration
@@ -1,1 +1,2 @@
+software.amazon.smithy.ruby.codegen.integrations.TestIntegration
 software.amazon.smithy.ruby.codegen.integrations.WhiteLabelTestIntegration
