--- codegen/smithy-ruby-codegen/src/main/resources/META-INF/services/software.amazon.smithy.ruby.codegen.RubyIntegration
@@ -1,1 +1,0 @@
-software.amazon.smithy.ruby.codegen.TestIntegration
